services/a_py/requirements.txt
grpcio==1.63.0
grpcio-tools==1.63.0
prometheus-client==0.20.0
